+++ src/pyannote/audio/core/pipeline.py
@@ -1,6 +1,7 @@
 # MIT License
 #
-# Copyright (c) 2021 CNRS
+# Copyright (c) 2021-2025 CNRS
+# Copyright (c) 2025- pyannoteAI
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -35,13 +36,13 @@ from pyannote.core.utils.helper import get_class_by_name
 from pyannote.database import FileFinder, ProtocolFile
 from pyannote.pipeline import Pipeline as _Pipeline
 
-from pyannote.audio import Audio, __version__
+from pyannote.audio import Audio
 from pyannote.audio.core.inference import BaseInference
 from pyannote.audio.core.io import AudioFile
 from pyannote.audio.core.model import Model
 from pyannote.audio.utils.hf_hub import AssetFileName, download_from_hf_hub
 from pyannote.audio.utils.reproducibility import fix_reproducibility
-from pyannote.audio.utils.version import check_version
+from pyannote.audio.utils.dependencies import check_dependencies
 
 
 def expand_subfolders(
@@ -156,14 +157,20 @@ class Pipeline(_Pipeline):
         with open(config_yml, "r") as fp:
             config = yaml.load(fp, Loader=yaml.SafeLoader)
 
+        # expand $model/{subfolder}-like entries in config
         expand_subfolders(
             config, model_id, revision=revision, token=token, cache_dir=cache_dir
         )
 
+        # before 4.x, pyannote.audio pipeline was using "version" key to
+        # specify the version of pyannote.audio used to train the pipeline
         if "version" in config:
-            check_version(
-                "pyannote.audio", config["version"], __version__, what="Pipeline"
-            )
+            config["dependencies"] = {"pyannote.audio": config["version"]}
+            del config["version"]
+
+        # check that dependencies are available (in their required version)
+        dependencies: dict[str, str] = config.get("dependencies", dict())
+        check_dependencies(dependencies, "Pipeline")
 
         # initialize pipeline
         pipeline_name = config["pipeline"]["name"]