+++ CHANGELOG.md
@@ -63,6 +63,7 @@ Models can now be stored alongside their pipelines in the same repository, strea
 - improve(train): speed up dataloaders
 - improve(setup): switch to `uv`
 - improve(setup): switch to `lightning` from `pytorch-lightning`
+- improve(utils): improve dependency check when loading pretrained models and/or pipeline
 
 ### Fixes